+++ SecurityComplianceCenter/internal/scripts/variables.ps1
@@ -1,1 +1,2 @@
-﻿
+﻿$script:PatternDisplayNameValidation = '[\\<>%&:;?/+|]'
+$script:PatternTooltipValidation = '[\x00\x08\x0B\x0C\x0E-\x1F]'